requirements.txt
streamlit
osmnx
networkx
folium
streamlit-folium==0.15.0
geopandas
pandas
numpy
scipy
reportlab
matplotlib
requests
scikit-learn
google-generativeai
fastapi
uvicorn
python-dotenv
python-multipart
anthropic
